Update_select-publishing-tool.rst_30-11-2017
I think it looks good for now. I suggest we update the website with the above text and then we contact all providers and ask them review and suggest updates if and when required.  We should  have a maximum of 200 words for the description of the tool/ service, information on whether it is a paid service and contacts details (which we already have for most providers).