Small suggestions and edits
Good work, Amy!!  The guidance is much clearer to what we had before- have a look at my edits and let me know if you are happy with them. We can then merge and update the website page :)

For changes of publisher id and Org IDs  we need to check whether we could automate this (maybe add to Registry list of features that we have), but for the current guidance we need to ask publishers to notify us.   Also, I think point 3 under ' Changing an organisation's publisher ID' can be made clearer- will check with you tomorrow .