Better indicate that E-to-NE changes happened to all versions
The change of some Codelists from Embedded to Non-Embedded happened with 2.03. It also impacts other versions.
As such, adding a couple of words to say that the change happened alongside 2.03 gives a reason for it to exist in the 2.03 changelog, while also indicating that it's part of a wider change.